kabi: add paravirt_patch_template to kABI
Patchwork-id: 46772
O-Subject: [RHEL 6.3 PATCH -v2 2/2] add paravirt_patch_template to kABI
Bugzilla: 813682
RH-Acked-by: Don Zickus <dzickus@redhat.com>

The kABI change to paravirt_patch_template fell through the
cracks because only structs that are function arguments can
be checked by our kABI tools.

To prevent this thing from breaking again in the future,
we need to add paravirt_patch_template to the kABI.
Unfortunately this seems to be the only way to do so.

If we have to extend paravirt_patch_template in the future,
we can do so only for functions used exclusively by the
core kernel, placing them at the end of the struct and
inside #ifndef __GENKSYMS__

A brew build is going here: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4368485

Fixes bug 813682

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>