nfs: Backport of upstream xdr routines to rhel6
Patchwork-id: 46989
O-Subject: [RHEL6.4 PATCH 27/27] RHEL6-only-patch:Backport of the upstream xdr routines to rhel6
Bugzilla: 784174
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

When RHEL6 branched from upstream, upstream decided to rewrite
the plumbing of NFS (aka the xdr routines of sunrpc).
We decided not to take that rewrite so this so this is a
backport of the upstream xdr routines to RHEL6.

Signed-off-by: Steve Dickson <steved@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>