mtip32xx: fix missing mtip32xx.ko in installer initrd
Patchwork-id: 47055
O-Subject: [RHEL6.3 PATCH] [v2]Fix missing mtip32xx.ko in modules.block
Bugzilla: 819947
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Jarod Wilson <jwilson@redhat.com>

From: Shyam Iyer <shiyer@redhat.com>

Kernel spec file generates modules.block and is missing the mtip32xx.ko driver in it. This is the case since the mtip32xx is not a regular block driver and the default symbols to match a block driver are not matched by the spec file.

This causes anaconda to not pick up the driver in the installer initramfs.
This can cause customer calls if we include driver support and installer initramfs doesn't contain the driver.

To fix this we need to pass a symbol to the spec file that matches in such a way that only the mtip32xx driver is added to the modules.block list.
 Eg: of these symbols used by the mtip32xx driver that are unique are...
dmam_alloc_coherent
pcim_iounmap_regions
dmam_free_coherent

However such symbols are irrelevant to match a block driver and run the risk of a new driver being pulled accidently in the future.

A somewhat safe symbol is blk_queue_physical_block_size which pulls in sd_mod.ko and zram.ko to modules.block in addition to mtip32xx.ko.
Jwilson says this should be safe as sd_mod.ko gets pulled in anyway and zram is harmless.

Tested by me. The modules.block file adds the mtip32xx.ko with this patch.

Thanks,
Shyam Iyer