mlx4: fix RoCE oops
Patchwork-id: 47094
O-Subject: [Patch RHEL6] ib/mlx4: take a hammer to the RoCE crash for now
Bugzilla: 749059
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

So, the fix to bug 749059 got kicked back out as FailsQA.  I found the
reason my original patch didn't work (update in the bug).  I still haven't
root caused why the OFED driver works and ours doesn't, but we need to
have something in place to stop this oops prior to 6.3 going out the
door.  So, with that in mind I give you my most recent changeset on the issue:

It's late in the game and I'd love to root cause this issue for sure,
but while I'm working on it, let's going something to solve the oops
in place.  This will prevent the oops and recover gracefully, it's just
not what OFED did and the OFED driver doesn't crash, so I'm still
looking for their magic bullet.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>