kstrto: Reduce the number of expensive division instructions done by _parse_integer()
Patchwork-id: 47122
O-Subject: [PATCH 8/8][RHEL6.4 BZ820847] Reduce the number of expensive division instructions done by _parse_integer()
Bugzilla: 820847
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Backport of the following upstream patch
--
commit 690d137f448d4c4da9001871e6569d5636f116c7
Author: David Howells <dhowells@redhat.com>
Date:   Thu Feb 9 15:48:20 2012 +0000

Reduce the number of expensive division instructions done by _parse_integer()

_parse_integer() does one or two division instructions (which are slow)
per digit parsed to perform the overflow check.

Furthermore, these are particularly expensive examples of division
instruction as the number of clock cycles required to complete them may
go up with the position of the most significant set bit in the dividend:

	if (*res > div_u64(ULLONG_MAX - val, base))

which is as maximal as possible.

Worse, on 32-bit arches, more than one of these division instructions
may be required per digit.

So, assuming we don't support a base of more than 16, skip the check if the
top nibble of the result is not set at this point.

Signed-off-by: David Howells <dhowells@redhat.com>
[ Changed it to not dereference the pointer all the time - even if the
  compiler can and does optimize it away, the code just looks cleaner.
  And edited the top nybble test slightly to make the code generated on
  x86-64 better in the loop - test against a hoisted constant instead of
  shifting and testing the result ]
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
--

Fixes BZ 820847

Signed-off-by: Sachin Prabhu
Signed-off-by: Jarod Wilson <jarod@redhat.com>