mqueue: add rbtree node caching support
Patchwork-id: 47221
O-Subject: [Patch RHEL6.4 1/2] ipc/mqueue: add rbtree node caching support
Bugzilla: 826177
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

When I wrote the first patch that added the rbtree support for message
queue insertion, it sped up the case where the queue was very full
drastically from the original code.  It, however, slowed down the
case where the queue was empty (not drastically though).

This patch caches the last freed rbtree node struct so we can quickly
reuse it when we get a new message.  This is the common path for any
queue that very frequently goes from 0 to 1 then back to 0 messages
in queue.

Andrew Morton didn't like that we were doing a GFP_ATOMIC allocation
in msg_insert, so this patch attempts to speculatively allocate a
new node struct outside of the spin lock when we know we need it, but
will still fall back to a GFP_ATOMIC allocation if it has to.

Once I added the caching, the necessary various ret = ; spin_unlock
gyrations in mq_timedsend were getting pretty ugly, so this also slightly
refactors that function to streamline the flow of the code and the
function exit.

Finally, while working on getting performance back I made sure that all
of the node structs were always fully initialized when they were first
used, rendering the use of kzalloc unnecessary and a waste of CPU cycles.

The net result of all of this is:

1) We will avoid a GFP_ATOMIC allocation when possible, but fall back on
it when necessary.
2) We will speculatively allocate a node struct using GFP_KERNEL if our
cache is empty (and save the struct to our cache if it's still empty after
we have obtained the spin lock).
3) The performance of the common queue empty case has significantly
improved and is now much more in line with the older performance for
this case.

The performance changes are:

            Old mqueue      new mqueue      new mqueue + caching
queue empty
send/recv   305/288ns       349/318ns       310/322ns

I don't think we'll ever be able to get the recv performance back, but
that's because the old recv performance was a direct result and consequence
of the old methods abysmal send performance.  The recv path simply must
do more so that the send path does not incur such a penalty under higher
queue depths.

As it turns out, the new caching code also sped up the various queue full
cases relative to my last patch.  That could be because of the difference
between the syscall path in 3.3.4-rc5 and 3.3.4-rc6, or because of the
change in code flow in the mq_timedsend routine.  Regardless, I'll take
it.  It wasn't huge, and I *would* say it was within the margin for error,
but after many repeated runs what I'm seeing is that the old numbers
trend slightly higher (about 10 to 20ns depending on which test is the
one running).

Signed-off-by: Doug Ledford <dledford@redhat.com>

Version 2:

Incorporate style changes from Andrew Morton
Don't bother passing new_leaf around...when I added the code inside
the spinlock to move new_leaf to info->node_cache, it really should
have been removed then as it now became impossible for us to ever
use new_leaf, we would either have a node_cache, or neither node_cache
or new_leaf.
Clean up the logic in msg_insert with this change in mind.

Performance also improved slightly with this logic change.  New
timings show a insert time of about 306ns.  That brings us right
in line with the old insert time.  Yay!

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>