be2iscsi: fix dma free size mismatch regression
Patchwork-id: 47303
O-Subject: [RHEL6.3 PATCH 1/1] be2iscsi: fix dma free size mismatch regression
Bugzilla: 824287
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 824287.

Patch has been posted here
http://marc.info/?l=linux-scsi&m=133782382718755&w=2

>From the post:
	As you can see from the patch I made a mistake. During
	development I switched from passing a struct to the size of
	the struct, but left the sizeof. This results in us allocating
	4 bytes (sizeof(int)) but then calling pci_free_consistent
	with the size of the struct.

I ran the patch on the box that we discovered the problem on,
but was not able to hit the trace/warnig before or after the patch. I also
ran the patch on my box here and did not see the problem. However,
it looks like a obvious bug and the reason for the trace/warning
in the bugzilla.

Signed-off-by: Jarod Wilson <jarod@redhat.com>