RHELPatchRules: Update BZ rules and add RH-AUTHOR
Patchwork-id: 47316
O-Subject: [RHEL6 PATCH INTERNAL] RHELPatchRules, Update BZ rules and add RH-AUTHOR
Bugzilla:
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

One request I keep seeing for this document is to allow users to state the
bugzilla number in either the subject OR the body, and drop the
requirement for stating it in both places.  This document update removes
the requirement for stating the bugzilla in both places, but restricts authors
to listing _one_ bugzilla in the subject and still allows for multiple
bugzillas to be specified in the body.

A second feature is the inclusion of the RH-AUTHOR field.  This will
allow submaintainers to submit patches in the name of other engineers so
those engineers will get credit for the patches.  This will minimize
cleanup work for the tree maintainers and the submaintainers.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>