Make keyctl new_session work on powerpc
Patchwork-id: 47388
O-Subject: [RHEL6.4 PATCH] Make keyctl new_session work on powerpc [BZ 830152]
Bugzilla: 830152
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

RH-Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=830152

"keyctl new_session" doesn't work on powerpc because the PowerPC arch's
TIF_NOTIFY_RESUME handler doesn't invoke the code to replace the session
keyring when KEYCTL_SESSION_TO_PARENT has been called.

Without this patch, the pending replacement session keyring is simply discarded
when the afflicted process exits or if further calls to the keyctl are made.
The replacement session keyring is not leaked.

With the patch, powerpc falls into line with the x86, ia64 and s390 arches.

Testing can be done by:

	keyctl add user a a @s
	keyctl show

This shows that the current session keyring now contains a user key called 'a'
and (presumably) a link the the per-UID keyring.  Doing:

	keyctl new_session

should then replace the session keyring of the shell from which it was run.
Running:

	keyctl show

again should now show an empty session keyring if it worked; if it didn't, then
it should show the same as the first time it was run.

The patch for BZ 827424 should also be applied - otherwise you risk a panic.

A working ppc64 kernel can be found here temporarily:

	https://brewweb.devel.redhat.com/taskinfo?taskID=4492023

Signed-off-by: David Howells <dhowells@redhat.com>