docs: Add RHELPatchRules-public
Patchwork-id: 4798
O-Subject: [kernel team] [RHEL6 PATCH INTERNAL] redhat/docs, Add RHELPatchRules-public
Bugzilla:

Requests have been made for an external version of the RHELPatchRules
document.  The existing RHELPatchRules document contains some RH internal
information that we do not wish to expose to partners or customers.

This version of the document is a sanitized version of the RHELPatchRules
document and can be provided to customers and partners who do not have
an onsite presence at Red Hat.  The document can also be provided to
onsite engineers to forward to their companies.

Signed-off-by: Jarod Wilson <jarod@redhat.com>