algapi: Move larval completion into algboss
Patchwork-id: 47779
O-Subject: Re: [RHEL6 PATCH] crypto: algapi - Move larval completion into algboss
Bugzilla: 832135
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

crypto: algapi - Move larval completion into algboss

RHEL6 Bugzilla #832135

It has been observed that sometimes the crypto allocation code
will get stuck for 60 seconds or multiples thereof.  This is
usually caused by an algorithm failing to pass the self-test.

If an algorithm fails to be constructed, we will immediately notify
all larval waiters.  However, if it succeeds in construction, but
then fails the self-test, we won't notify anyone at all.

This patch fixes this by merging the notification in the case
where the algorithm fails to be constructed with that of the
the case where it pases the self-test.  This way regardless of
what happens, we'll give the larval waiters an answer.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jarod Wilson <jarod@redhat.com>