perf: Fix build dependency of perf python extension
Patchwork-id: 47993
O-Subject: [PATCH RHEL6.4 BZ827474 027/122] perf tools: Fix build dependency of perf python extension
Bugzilla: 827474
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 827474
https://bugzilla.redhat.com/show_bug.cgi?id=827474

upstream status
===============
commit 6a5c13aff49ac9b3fea38d5f84b436718cb2780d
Author: Namhyung Kim <namhyung@gmail.com>
Date:   Sun Feb 12 19:45:24 2012 +0900

description
===========
The perf python extention (perf.so) file lacks its dependencies in the
Makefile so that it cannot be refreshed if one of source files it depends
is changed. Fix it by putting them in a separate file and processing it in
both of Makefile and setup.py.