perf: fix broken perf record -a mode
Patchwork-id: 48016
O-Subject: [PATCH RHEL6.4 BZ827474 045/122] perf tools: fix broken perf record -a mode
Bugzilla: 827474
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 827474
https://bugzilla.redhat.com/show_bug.cgi?id=827474

upstream status
===============
commit 6b1bee9035d430c4b4f586df6df4b3f840e89b5b
Author: Stephane Eranian <eranian@google.com>
Date:   Tue Feb 21 15:54:25 2012 +0100

description
===========
The following commit:
b52956c perf tools: Allow multiple threads or processes in record, stat, top

introduced a bug in the thread_map code which caused perf record -a to
not setup system-wide monitoring properly.

$ taskset -c 1 noploop 1000 &
$ perf record -a -C 1 sleep 10
$ perf report -D | tail -20
cycles stats:
           TOTAL events:       4413
            MMAP events:       4025
            COMM events:        340
          SAMPLE events:         48

Here I was expecting about 10,000 samples and not 48.

In system-wide mode, the PID passed to perf_event_open() must be -1 and
it was 0. That caused the kernel to setup a per-process event on PID:0.
Consequently, the number of samples captured does not correspond to the
requested measurement.

The following one-liner fixes the problem for me with or without -C.

I would also suggest to change the malloc() to something that matches
the struct definition. thread_map->map[] is declared as int map[] and
not pid_t map[]. If map[] can only contain pids, then change the struct
definition.