perf: Incorrect use of snprintf results in SEGV
Patchwork-id: 48062
O-Subject: [PATCH RHEL6.4 BZ827474 085/122] perf tools: Incorrect use of snprintf results in SEGV
Bugzilla: 827474
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 827474
https://bugzilla.redhat.com/show_bug.cgi?id=827474

upstream status
===============
commit b832796caa1fda8516464a003c8c7cc547bc20c2
Author: Anton Blanchard <anton@samba.org>
Date:   Wed Mar 7 11:42:49 2012 +1100

description
===========
I have a workload where perf top scribbles over the stack and we SEGV.
What makes it interesting is that an snprintf is causing this.

The workload is a c++ gem that has method names over 3000 characters
long, but snprintf is designed to avoid overrunning buffers. So what
went wrong?

The problem is we assume snprintf returns the number of characters
written:

    ret += repsep_snprintf(bf + ret, size - ret, "[%c] ", self->level);
...
    ret += repsep_snprintf(bf + ret, size - ret, "%s", self->ms.sym->name);

Unfortunately this is not how snprintf works. snprintf returns the
number of characters that would have been written if there was enough
space. In the above case, if the first snprintf returns a value larger
than size, we pass a negative size into the second snprintf and happily
scribble over the stack. If you have 3000 character c++ methods thats a
lot of stack to trample.

This patch fixes repsep_snprintf by clamping the value at size - 1 which
is the maximum snprintf can write before adding the NULL terminator.

I get the sinking feeling that there are a lot of other uses of snprintf
that have this same bug, we should audit them all.