perf: Fix bug in raw sample parsing
Patchwork-id: 48051
O-Subject: [PATCH RHEL6.4 BZ827474 092/122] perf tools: Fix bug in raw sample parsing
Bugzilla: 827474
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 827474
https://bugzilla.redhat.com/show_bug.cgi?id=827474

upstream status
===============
commit fa30c964c0ac31a705a4f86072267a42343c4158
Author: Stephane Eranian <eranian@google.com>
Date:   Sat Mar 17 23:23:18 2012 +0100

description
===========
In perf_event__parse_sample(), the array variable was not incremented
by the amount of data used by the raw_data.

That was okay until we added PERF_SAMPLE_BRANCH_STACK which depends on
the array variable pointing to the beginning of the branch stack data.

But that was not the case if branch stack was combined with raw mode
sampling. That led to bogus branch stack addresses and count.

The bug would show up with:
$ perf record -R -b foo

This patch fixes the problem by correctly moving the array pointer
forward for RAW samples.