perf, stat: Fix case where guest/host monitoring is not supported by kernel
Patchwork-id: 48077
O-Subject: [PATCH RHEL6.4 BZ827474 116/122] perf stat: Fix case where guest/host monitoring is not supported by kernel
Bugzilla: 827474
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 827474
https://bugzilla.redhat.com/show_bug.cgi?id=827474

upstream status
===============
commit 5622c07b4741e0afd7607bce6e850b76eeb23210
Author: Stephane Eranian <eranian@google.com>
Date:   Fri Apr 27 14:45:38 2012 +0200

description
===========
By default, perf stat sets exclude_guest = 1. But when you run perf on a
kernel which does not support  host/guest filtering, then you get an
error saying the event in unsupported. This comes from the fact that
when the perf_event_attr struct passed by the user is larger than the
one known to the kernel there is safety check which ensures that all
unknown bits are zero. But here, exclude_guest is 1 (part of the unknown
bits) and thus the perf_event_open() syscall return EINVAL.

To my surprise, running perf record on the same kernel did not exhibit
the problem. The reason is that perf record handles the problem by
catching the error and retrying with guest/host excludes set to zero.
For some reason, this was not done with perf stat. This patch fixes this
problem.