fs/direct-io.c: dont try to allocate more than BIO_MAX_PAGES in a bio
Patchwork-id: 48540
O-Subject: [RHEL6 PATCH] fs/direct-io.c: don't try to allocate more than BIO_MAX_PAGES in a bio
Bugzilla: 832962
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

Hi,

This is a backport of the following:

commit 20d9600cb407b0b55fef6ee814b60345c6f58264
Author: David Dillow <dillowda@ornl.gov>
Date:   Thu Jan 20 14:44:22 2011 -0800

    fs/direct-io.c: don't try to allocate more than BIO_MAX_PAGES in a bio

    When using devices that support max_segments > BIO_MAX_PAGES (256), direct
    IO tries to allocate a bio with more pages than allowed, which leads to an
    oops in dio_bio_alloc().  Clamp the request to the supported maximum, and
    change dio_bio_alloc() to reflect that bio_alloc() will always return a
    bio when called with __GFP_WAIT and a valid number of vectors.

    [akpm@linux-foundation.org: remove redundant BUG_ON()]
    Signed-off-by: David Dillow <dillowda@ornl.gov>
    Reviewed-by: Jeff Moyer <jmoyer@redhat.com>
    Cc: <stable@kernel.org>
    Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
    Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Our customers/partners have hit this when using iSCSI, servicing I/Os of
anything larger than 1MB.  The fix is pretty simple, just capping the
number of pages asked for the the maximum allowed in a bio.

The upstream patch applied without any conflicts or fuzz.  I built it
locally (not in brew), and did not do targeted testing, but the fix is
really straight-forward.  This resolves bug 832962:
  https://bugzilla.redhat.com/show_bug.cgi?id=832962

Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Jarod Wilson <jarod@redhat.com>