openvswitch: Add Open vSwitch kernel components
Patchwork-id: 48961
O-Subject: [PATCH 11/20] net: Add Open vSwitch kernel components.
Bugzilla: 805656
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>

RHEL6 Bugzilla #805656

Upstream commit: ccb1352e76cff0524e7ccb2074826a092dd13016

Open vSwitch is a multilayer Ethernet switch targeted at virtualized
environments.  In addition to supporting a variety of features
expected in a traditional hardware switch, it enables fine-grained
programmatic extension and flow-based control of the network.
This control is useful in a wide variety of applications but is
particularly important in multi-server virtualization deployments,
which are often characterized by highly dynamic endpoints and the need
to maintain logical abstractions for multiple tenants.

The Open vSwitch datapath provides an in-kernel fast path for packet
forwarding.  It is complemented by a userspace daemon, ovs-vswitchd,
which is able to accept configuration from a variety of sources and
translate it into packet processing rules.

See http://openvswitch.org for more information and userspace
utilities.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>