openvswitch: Add length check when retrieving TCP flags
Patchwork-id: 48955
O-Subject: [PATCH 16/20] openvswitch: Add length check when retrieving TCP flags.
Bugzilla: 805656
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>

RHEL6 Bugzilla #805656

Upstream commit: bf32fecdc1851ad9ca960f56771b798d17c26cf1

When collecting TCP flags we check that the IP header indicates that
a TCP header is present but not that the packet is actually long
enough to contain the header.  This adds a check to prevent reading
off the end of the packet.

In practice, this is only likely to result in reading of bad data and
not a crash due to the presence of struct skb_shared_info at the end
of the packet.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>