cifs: fix handling of scopeid in cifs_convert_address
Patchwork-id: 49028
O-Subject: [PATCH] [bz823625] cifs: fix handling of scopeid in cifs_convert_address
Bugzilla: 823625
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Fixes bz 823625

Backport of the following upstream patches
--
commit 9616125611ee47693186533d76e403856a36b3c8
Author: Jeff Layton <jlayton@redhat.com>
Date:   Wed Feb 16 09:34:16 2011 -0500

cifs: fix handling of scopeid in cifs_convert_address

The code finds, the '%' sign in an ipv6 address and copies that to a
buffer allocated on the stack. It then ignores that buffer, and passes
'pct' to simple_strtoul(), which doesn't work right because we're
comparing 'endp' against a completely different string.

Fix it by passing the correct pointer. While we're at it, this is a
good candidate for conversion to strict_strtoul as well.

Cc: stable@kernel.org
Cc: David Howells <dhowells@redhat.com>
Reported-by: BjÃ¶rn JACKE <bj@sernet.de>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
--

--
commit 2545e0720a5a4bf8ebccc6f793f97a246cf3f18d
Author: Dan Carpenter <dan.carpenter@oracle.com>
Date:   Thu Mar 1 10:06:52 2012 +0300

cifs: writing past end of struct in cifs_convert_address()

"s6->sin6_scope_id" is an int bits but strict_strtoul() writes a
long
so this can corrupt memory on 64 bit systems.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
--

Signed-off-by: Sachin Prabhu
Signed-off-by: Jarod Wilson <jarod@redhat.com>