cifs: Allow to set extended attribute cifs_acl
Patchwork-id: 49064
O-Subject: [PATCH 02/11] [bz824065] Allow to set extended attribute cifs_acl
Bugzilla: 824065
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Fixes bz 824065

Backport of the following patch
--
commit b73b9a4ba753dfd7d304ee6ee4685b827524c533
Author: Steve French <sfrench@us.ibm.com>
Date:   Tue Apr 19 18:27:10 2011 +0000

[CIFS] Allow to set extended attribute cifs_acl (try #2)

Allow setting cifs_acl on the server.
Pass on to the server the ACL blob generated by an application.
cifs is just a pass-through, it does not monitor or inspect the contents
of the blob, server decides whether to enforce/apply the ACL blob composed
by an application.
If setting of ACL is succeessful, mark the inode for revalidation.

Signed-off-by: Shirish Pargaonkar <shirishpargaonkar@gmail.com>
Acked-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
--

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>