cifs: Make cifsFileInfo_put work with strict cache mode
Patchwork-id: 49093
O-Subject: [RHEL6 PATCH 0.5/7] [bz742314] CIFS: Make cifsFileInfo_put work with strict cache mode
Bugzilla: 742314
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

I missed out sending this patch along with the other patches. This is
patch number 1 of the total 8 patches sent for the strictcache mount
option.

Fixes bz 742314

Backport of upstream commit
--
commit 4f8ba8a0c095933dd54a2c281750c8a85b329b26
Author: Pavel Shilovsky <piastryyy@gmail.com>
Date:   Sun Nov 21 22:36:12 2010 +0300

CIFS: Make cifsFileInfo_put work with strict cache mode

On strict cache mode when we close the last file handle of the inode we
should set invalid_mapping flag on this inode to prevent data coherency
problem when we open it again but it has been modified on the server.

Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
--

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>